pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "automeister"
version = "0.1.0"
description = "Desktop automation tool for Linux X11 environments"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Automeister Contributors" }
]
keywords = ["automation", "desktop", "x11", "xdotool"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Environment :: X11 Applications",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Desktop Environment",
    "Topic :: Utilities",
]
dependencies = [
    "typer[all]>=0.9.0",
    "pyyaml>=6.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
automeister = "automeister.cli:app"

[project.urls]
Homepage = "https://github.com/automeister/automeister"
Documentation = "https://github.com/automeister/automeister#readme"
Repository = "https://github.com/automeister/automeister"

[tool.hatch.build.targets.wheel]
packages = ["src/automeister"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
